Update migrator to use the data_store_version library
