+++ workspaces/api/migration/migrator/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 publish = false
 
 [dependencies]
+data_store_version = { path = "../../data_store_version" }
 fs_extra = "1.1"
 lazy_static = "1.2"
 log = "0.4"
