+++ workspaces/api/migration/migrator/src/args.rs
@@ -1,6 +1,6 @@
 //! This module handles argument parsing for the migrator binary.
 
-use crate::version::Version;
+use data_store_version::Version;
 use std::env;
 use std::fs;
 use std::path::PathBuf;
